refactor: split utils, moving funcs to common, fix problem with icons
