+++ src/view/edit-event.js
@@ -1,4 +1,4 @@
-import { formatDate } from '../utils/utils';
+import { formatDate } from '../utils/date';
 import { Types, Destinations } from '../const';
 import { createPhotoContainerTemplate } from './pictures';
 import { createOffersContainerTemplate } from './offers';
@@ -38,7 +38,6 @@ export const createEditEventTemplate = (eventItem) => {
     : createCheckedTypeItemTemplate(eventType)).join('');
 
   const destinations = Object.values(Destinations).map(createOptionTemplate).join('');
-  // console.log(destinations);
 
   return `<li class="trip-events__item">
     <form class="event event--edit" action="#" method="post">
@@ -46,7 +45,7 @@ export const createEditEventTemplate = (eventItem) => {
         <div class="event__type-wrapper">
           <label class="event__type  event__type-btn" for="event-type-toggle-1">
             <span class="visually-hidden">Choose event type</span>
-            <img class="event__type-icon" width="17" height="17" src="img/icons/${type}.png" alt="Event type icon">
+            <img class="event__type-icon" width="17" height="17" src="img/icons/${type.toLowerCase()}.png" alt="Event type icon">
           </label>
           <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
