+++ src/view/event.js
@@ -1,4 +1,4 @@
-import { getDate, getStart, getEnd, getGap, gapToString } from '../utils/utils';
+import { getDate, getStart, getEnd, getGap, gapToString } from '../utils/date';
 
 const createOfferTemplate = (offer) => (
   `<li class="event__offer">
@@ -27,7 +27,7 @@ export const createEventTemplate = (eventItem) => {
     <div class="event">
       <time class="event__date" datetime="2019-03-18">${getDate(dateFrom)}</time>
       <div class="event__type">
-        <img class="event__type-icon" width="42" height="42" src="img/icons/${type}.png" alt="Event type icon">
+        <img class="event__type-icon" width="42" height="42" src="img/icons/${type.toLowerCase()}.png" alt="Event type icon">
       </div>
       <h3 class="event__title">${type} ${destination.name}</h3>
       <div class="event__schedule">
