+++ src/view/offers.js
@@ -1,19 +1,12 @@
 const createOfferTemplate = (offer) => (
-  offer.isChecked ? `<div class="event__offer-selector">
-      <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" checked>
+  `<div class="event__offer-selector">
+      <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" ${offer.isChecked ? 'checked' : ''}>
       <label class="event__offer-label" for="event-offer-comfort-1">
         <span class="event__offer-title">${offer.title}</span>
         &plus;&euro;&nbsp;
         <span class="event__offer-price">${offer.price}</span>
       </label>
-    </div>` : `<div class="event__offer-selector">
-    <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort">
-    <label class="event__offer-label" for="event-offer-comfort-1">
-      <span class="event__offer-title">${offer.title}</span>
-      &plus;&euro;&nbsp;
-      <span class="event__offer-price">${offer.price}</span>
-    </label>
-  </div>`
+    </div>`
 );
 
 export const createOffersContainerTemplate = (offers) => (
